fix: 11 separate all authentication logic to other app (#16)
