+++ dataset_retriever/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for retriever project.
+ASGI config for dataset_retriever project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'retriever.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'dataset_retriever.settings')
 
 application = get_asgi_application()
