+++ dataset_retriever/celery.py
@@ -4,9 +4,9 @@ from celery import Celery
 from celery.schedules import crontab
 
 # Set the default Django settings module for the 'celery' program.
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'retriever.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'dataset_retriever.settings')
 
-app = Celery('retriever')    
+app = Celery('dataset_retriever')    
 
 # Using a string here means the worker doesn't have to serialize
 # the configuration object to child processes.
